requirements.txt
Flask
torch
torchvision
Pillow
opencv-python
numpy